requirements.txt
fastapi
uvicorn
pymongo
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-fastapi
pytest
pytest-asyncio
httpx
python-dotenv
kafka-python
rich